DETAIL PAGE 댓글 CSS 추가

+++ Acorn/src/main/java/com/gura/acorn/shop/dto/ShopReviewDto.java
@@ -11,7 +11,7 @@ public class ShopReviewDto {
 	private int ref_group;
 	private int review_group;
 	private String deleted;
-	private int grade; //필요 시 다른테이블 dto로 이동
+	private double grade; //필요 시 다른테이블 dto로 이동
 	private String regdate;
 	private String profile;
 	private int startRowNum;
@@ -21,7 +21,7 @@ public class ShopReviewDto {
 	public ShopReviewDto() {}
 
 	public ShopReviewDto(int num, String writer, String content, String title, int ref_group, int review_group,
-			String deleted, int grade, String regdate, String profile, int startRowNum, int endRowNum,
+			String deleted, double grade, String regdate, String profile, int startRowNum, int endRowNum,
 			String imagePath) {
 		super();
 		this.num = num;
@@ -95,11 +95,11 @@ public class ShopReviewDto {
 		this.deleted = deleted;
 	}
 
-	public int getGrade() {
+	public double getGrade() {
 		return grade;
 	}
 
-	public void setGrade(int grade) {
+	public void setGrade(double grade) {
 		this.grade = grade;
 	}
 
@@ -143,6 +143,6 @@ public class ShopReviewDto {
 		this.imagePath = imagePath;
 	}
 
-
+	
 	
 }
